requirements.txt
GitPython
loguru
PyGithub==2.1.1
